requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
qrcode[pil]
Pillow
pytz